Init alsa everytime heaphone plugged in
After system upgrade, headphones are not automatically recognized. It
seems the machine directs the signal to headphones, but sound doesn't
come out. `alsactl init` resolves this, but it should be called after
headphones are plugged in. ACPI handler handles this.

Note that calling amixer to get a volumn value with sudo privilege is
different from that with user privilege, probably because of
pulseaudio-alsa layer. So it may not recover current volume level
properly.